Enable running multiple autopilot instances (#3580)
# Description

Enable multiple autopilot instances to run at the same time by one
instance be the leaders doing writes all other instances being in stand
by mode keeping their caches warm and waiting for the leader to step
down.

**The problem**
Autopilot takes its sweet time to start up, it can be a more than a
minute before it populates its caches and actually start cutting new
auctions. At the same time the autopilot is not meant to have multiple
instances running at the same time, should we do that weird things might
happen. In addition the autopilot doesn't have a /ready endpoint that
k8s could use in a rolling deployment, there is no way to know when a
new instance might be ready.

**The solution**
With some adjustments we could have two autopilot modes: a "leader" mode
that does everything the autopilot does today and a "follower" mode that
keeps its caches/state warm, but doesn't actually cut auctions. These
can run side by side and when the leader dies the stand by becomes the
leader. In a deployment we increase the number of instances by one, the
new instance sees the leader spot is occupied and run in stand by mode.
When it's ready (indicated by 200 on /ready) k8s kills the old autopilot
instance(s) and the new one becomes the leader.

**Implementation** / Changes
- [x] Introduce a postgres lock that the leader holds to do "write" work
and instances without it are on standby ready to jump in when the
leaders steps down
- [x] Extract read only parts that warm up caches
- [x] Introduce a /ready endpoint that returns 200 only when the
instance is ready to start cutting auctions (warmed up caches)
- [x] Add a sigint/signterm handler to 1) finish the current auction and
exit 2) drop the pg lock
- [x] Write an integration test 
- [x] Configure k8s to use a rolling deployment where we start a new
autopilot, wait for it to become /ready and send sigterm to the old
leader instance
- [x] Decision: Decide whether we should have 2 instances running at all
times for stuff like OOM kills or just during deployment (constraint
here might be fetching native prices)


## How to test
<!--- Include details of how to test your changes, including any
pre-requisites. If no unit tests are included, please explain why and
how to test manually
1.
2.
3.
-->

<!--
## Related Issues

Fixes #
-->

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>
Co-authored-by: Marcin Szymczak <mail@marszy.com>
Co-authored-by: ilya <ilya@cow.fi>
Co-authored-by: Bram van den Berg <bram@cow.fi>
Co-authored-by: José Duarte <duarte.gmj@gmail.com>
Co-authored-by: Nuno Anselmo <me@nunoanselmo.me>