Limit pool ids in `where` clause when querying univ3 subgraph for Ticks (#3632)
Some slow nodes time out when there are too many pool IDs when we query
the uniswap v3 subgraph for `Tick`s. This PR adds the option to limit
the number of such items in the query which means running multiple
queries with fewer pool IDs in `where` to achieve the same effect. This
can't be done in parallel as it overloads these slow nodes.

In the end I found a [fast
graph](https://thegraph.com/explorer/subgraphs/FiJDXMFCBv88GP17g2TtPh8BcA8jZozn5WRW7hCN7cUT?view=About&chain=arbitrum-one)
so this change is not needed, but if this one ever stops working we
might be happy to have the option.

- [x] Add a new driver param that limits how many pool ids are added to
a single query

Start driver with graph-url against a slow subgraph (e.g.
`F85MNzUGYqgSHSHRGgeVMNsdnW1KtZSVgFULumXRZTw2`) and watch it fail. Then
set max pool ids to 10 and see it succeed.

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>