Speed up settlement indexing (#3654)
# Description
Currently the autopilot's maintenance tasks that it runs between 2
auctions is very slow. The biggest time sink is currently indexing new
settlement events. After some investigation it became apparent that the
DB query we do to get all the necessary data for post-processing the
auction is very slow (250ms - 500ms).
There are 2 reasons for that:
1. we fetch all the data from the given auction - not just what we
actually need
2. we do multiple DB queries instead of doing 1 big query

<img width="1189" height="343" alt="Screenshot 2025-09-12 at 13 41 17"
src="https://github.com/user-attachments/assets/c3078575-16a3-4fa4-8b9a-72b62e9f5a96"
/>
<img width="1189" height="300" alt="Screenshot 2025-09-12 at 13 41 54"
src="https://github.com/user-attachments/assets/9d5bd008-c88a-49ee-b1af-9b5425f08c95"
/>

Because this is a very important query I didn't go ahead and tested the
performance in the actual cluster yet. ☝️ is just the status quo.


# Changes
This PR addresses both of these issues. So far I only tested it with a
few hand crafted test queries but they were in the ball bark of 1-2ms
execution time.
Because we return data with a lot of different cardinalities and
structures I decided to return the data as a JSON string and parse it
into our domain object. That made the query a bit difficult to read but
I think the performance improvement should still be worth it.

## How to test
Given that this query if fundamental for indexing events which almost
all of the e2e tests have some form of assertion on I think running the
e2e tests in CI should be sufficient for the correctness.