Adjust hooks counter test (#3767)
# Description
As it [was
suggested](https://github.com/cowprotocol/services/pull/3762#pullrequestreview-3320393408)
in the previous PR:
> Fine to merge as is but I just realized that this test is actually
just verifying that both hooks get executed but not that they get
executed but not WHEN they get executed.
> In order to verify that we could adjust the Counter to set the count
to the user's sell token balance at the time of executing the hook.
Then we'd assert that the pre counter has the value of the user's sell
token balance before the settlement starts and post has the value of
afterwards. Since the value the count is set to depends on data that
changes throughout the settlement we'd know when the hook was actually
executed.

# Changes

- Update the test Counter helper so hooks can overwrite a counter with a
live ERC20 balance snapshot.
- Adjust the Counter ABI JSON to add the new function while leaving the
original structure intact. To better understand the change, use this
diff:
https://github.com/cowprotocol/services/pull/3767/commits/9cf8761c8bcc59d5c762ff750d3b5070dfb34c38
- Rework the partial-fills hook e2e test to assert pre/post hook
execution timing by checking WETH balances recorded by the new helper.

---------

Co-authored-by: José Duarte <duarte.gmj@gmail.com>