CoW AMM indexer DB storage (#3684)
# Description
Currently, on each service restart, CoW AMM indexing starts from the AMM
Factory SC's deployment block, since all the indexed events are stored
in in-memory cache. This has become problematic on some chains that
actively use AMMs, since on each restart, the service has to re-index
all the events. As a result, this consumes a lot of RPC resources, and
the overall protocol performance drops drastically. On Base, for
example, it might take 2-3h.

# Changes
This PR introduces a persistence layer for the CoW AMM indexing, which
effectively stores all the indexed AMMs and the last indexed block, so
on the next restart, it can continue from the last indexed block.

- [ ] When the registry is initialized, fetch CoW AMMs from the DB to
populate the cache. Since we have an indexer per factory, the factory
address is stored in the db.
- [ ] The `last_indexed_blocks` will now contain information per CoW AMM
factory, which is required for initialization of each indexer. If no
data is found, use the configured deployment block.
- [ ] When another CoW AMM is indexed, it is stored in both in-memory
cache and the db.
- [ ] On revert, all the CoW AMMs with the affected blocks are removed.

## How to test
Existing tests. Updated postgres tests. Staging.