Index ethflow orders in the background (#3849)
# Description
This moves ethflow order indexing from the critical path between 2
auctions to a background task.
The reason is an insane influx of ethflow orders that significantly slow
down the auction process at the moment.
This is primarily intended as a bandaid.

# Changes
ethflow indexing now happens in a background task

## How to test
updated the ethflow specific tests to continuously mint new blocks since
the indexing of new ethflow orders is no longer synced to building
auctions.