[TRIVIAL] Persist cow amm factory address instead of helper in DB (#3860)
# Description
Currently we store the cow amm's helper contract as its factory address.
But when we
[load](https://github.com/cowprotocol/services/blob/main/crates/cow-amm/src/cache.rs#L56)
them from the DB to init the cache we search for them by the factory
address.

This was not caught by tests because we don't have a test that:
1. indexes new cow amms
2. shuts down the autopilot
3. restarts the autopilot
4. inits the autopilot's cow amm cache from the DB

# Changes
This PR fixes the issue by correctly using the factory address for
storing and loading.

I fixed the incorrect DB entries and after restarting the autopilot pods
all cow AMMs got loaded correctly.