Make tx gas limit configurable in driver (#3780)
Required to fix Sepolia after fusaka hard fork.
Sepolia got broken because the fusaka hard fork introduced a protocol
level cap on tx gas limit which currently is being set to the block gas
limit (which is too high).
Make tx gas limit configurable in the driver, preserving the old
behaviour (of taking block gas limit) if it is not specified.
- [ ] Add tx_gas_limit command line argument (env: TX_GAS_LIMIT)
1. Configure limit to 2^24 - 1 and test if transactions can be made on
Sepolia
2.

Fixes #3777