Migrate authenticator contract to alloy (#3817) Migrates the authenticator contract (source of truth for which addresses are allow listed solvers) to use alloy. - [x] removed old bindings - [x] added new bindings - [x] fixed type errors until the compiler stopped yelling at me
I also added `ProviderSignerExt::without_wallet()` to get a fresh
provider that can be used to sign impersonated accounts. That is needed
because telling `anvil` to impersonate some address does not magically
add the signer to the wallet. So when we have a wallet configured and
ask to sign for a particular address `alloy` complains that it doesn't
know how to sign. If we don't have any wallet configured alloy will just
forward the tx to the node for signing.
- compiler
- existing e2e tests