feat: generalized wrappers (#3700)
this introduces the ability for a `wrapper` contract to be included in
the appdata and solutions. A `wrapper` is a contract which is called
with the same parameters as settlement contract `settle`, but executes a
series of operations either/both before or after executing the actual
GPv2Settlement contract's `settle` function. This simplifies
integration. see the full info here
https://www.notion.so/cownation/Generalized-Wrapper-2798da5f04ca8095a2d4c56b9d17134e

Only approved `wrapper` contracts will be allowed to call the
`GPv2Settlement` contract (via the current GPv2AllowlistAuthenticator),
and it is expected that wrapper contracts will do the necessary
validation of solvers on behalf of the settlement contract.

Its expected that generalized wrappers can simplify the implementation
of many initiatives, including:
* "flashloans" integrations (ex. aave, euler)
* enforcable pre- and post-hooks (users can ensure that post actions
must at least be attempted, such as deposit to a vault)
* ex: with our current CowShed implementation, if cross chain send were
to fail due to ex. insufficient fee, funds can be sent back to the user
rather than needing to be recovered from the CowShed contract
* TWAP
* currently we are implementing this for EOAs, but its very difficult to
ensure funds are only pulled from the users wallet when needed because
there is no auth. Since the CowShed could effectively become a wrapper
contract, it can pull funds from user and immediately call the necessary
`settle` function afterwards to ensure funds are traded.

* `wrappers` is added to the `appData` so that the frontend can "hint"
the necessary flashloan or other required functionality if needed
(similar to current `flashloans` functionality). More than one wrapper
can be specified in a single order. A wrapper is combination of
`address`, the address of the wrapper contract to execute, and `data`,
which is any additional data required for the wrapper to execute
* `wrappers` is added to solution so its easy for the driver to encode
into a final solution. It is an array in order to accomodate different
orders that require different wrappers.
* additionally a solver can add their own wrapper if they determine its
necessary/useful to get the best result from a trade (some already
technically do this)

Potentially further changes required/limitations:
* when quoting, how can the use of a wrapper be effectively included in
the calculation?
* this implements handling in the autopilot and driver, but what will
need to happen from the solver side?

I have deployed a testing `EmptyWrapper` to
`0x54112E2F481AC239661914691082039d7B05A264`. By using this contract as
a simple wrapper, created a script to serve as a minimal E2E POC:

This PR can be verified to be working E2E by taking the following steps:
1. Boot up the playground environment. I used the command: `docker
compose -f playground/docker-compose.fork.yml up --build`
2. With all the services stabilized and running, open a new terminal and
run the provided wrapper test script: `./playground/test_wrapper.sh`
3. In the docker compose logs, observe the order being placed, an
auction held, and finally a transaction submitted to chain
4. The transaction result can be digested with `cast`:
* To get the submitted transaction and all its data fields: `cast tx <my
transaction hash from logs>`
* To get the result of executing the transaction on-chain (the receipt):
`cast receipt <my transaction hash from logs>`
5. It is also possible to see the order information and the active
fields for `wrapper` and `wrapperData` by opening up
https://localhost:8001 (the CoW Explorer) and putting in the printed
order ID at the end of script execution.
* general wiki
https://www.notion.so/cownation/Generalized-Wrapper-2798da5f04ca8095a2d4c56b9d17134e
* PR on the contract side
https://github.com/cowprotocol/euler-integration-contracts/pull/6

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>