Use `FuturesUnordered` when fetching COW AMM balances (#3924)
# Description
Fetching COW AMM token balances takes a lot of time on base due to a
high number of AMMs. In order to speed things up a bit, it is suggested
to use `FuturesUnordered`, which is a better approach than
`FuturesOrdered`, which is used by `futures_util::future::join_all` when
the number of futures exceeds 30.

## How to test
Staging