Post process settlements in parallel (#3853)
# Description
Currently we post-process settlements (i.e. associate the tx with a
solution proposed by a solver) serially. IIRC it was done that way
because it was simpler and that step simply ran in a background task at
the time.
Since then this logic has been moved onto the hot path so all the time
we spend there delays the creation of the next auction.
Additionally with the introduction of combinatorial auctions it's now
possible and relatively common to have multiple settlement transactions
in the same block. Whenever we have multiple settlements in one block
this PR should result in a significant performance uplift because we
process multiple concurrently.

# Changes
replaced loop to post-process auctions serially with logic that first
fetches ALL unprocessed settlements and then works on up to 10 of them
concurrently. This is fine for the post processing logic (as opposed to
the raw event indexing) because it's okay for post-processing to happen
out of order. So if the we successfully post-process settlement `n+1`
but not settlement `n` (e.g. due to a crash) the DB query would still
just return all unprocessed events instead of all NEW unprocessed
events.

I also adjusted how retrying works in this code. Instead of returning
some result enum (ok, invalid, nothing_to_do) to make the serial loop
retry something we now have a function `retry_with_sleep` that simply
retires a passed in future n times and returns an `Option` to indicate
whether it was successful or not.
This is used in 2 places:
1. fetching all outstanding settlements
2. post-processing individual outstanding settlement

With that the new logic should be as robust as the old one while being
IMO easier to reason about.

## How to test
Not sure how to test the improvement specifically. Since this is a
performance optimization and I don't really want to test the internals
of the implementation having a new test that makes sure that a big
number of settlements can be post-processed would be enough to test
correctness and for the performance aspect we'd have to deploy this on
the cluster.

Regarding the performance I temporarily deployed it to staging and it
produced the expected effect of reducing the spikiness that comes from
multiple settlements needing to be post-processed in the same block.
<img width="954" height="270" alt="Screenshot 2025-11-03 at 22 12 26"
src="https://github.com/user-attachments/assets/b459fb76-bf44-40db-894e-46ffd3d67f7d"
/>